restructuring from features to view
