+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { default as CocktailCard } from "./CocktailCard";
+export { default as Warning } from "./Warning";
+
+// layout
+
+export { default as CocktailSection } from "./layout/CocktailSection";
