--- src/features/cocktails/index.css
@@ -1,28 +1,0 @@
-.cocktail-section {
-  max-width: 700px;
-  margin: 0 auto;
-}
-
-.cocktail-grid {
-  width: 100%;
-  margin: 0 auto;
-  display: grid;
-  row-gap: 44px;
-}
-
-.cocktail-grid-card img {
-  width: 95%;
-  height: auto;
-  object-fit: cover;
-}
-
-@media screen and (min-width: 767px) {
-  .cocktail-grid {
-    width: max-content;
-    grid-template-columns: repeat(4, 158px);
-    grid-template-rows: repeat(2, 190px);
-    column-gap: 37px;
-    row-gap: 23px;
-  }
-  
-}
