--- src/features/generalTypes.ts
@@ -1,3 +1,0 @@
-export type ApiResponseType = {
-  [key: string]: string | number | boolean;
-};
