+++ src/store/index.ts
@@ -1,0 +1,25 @@
+import {
+  legacy_createStore as createStore,
+  combineReducers,
+  applyMiddleware,
+} from "redux";
+import { composeWithDevTools } from "redux-devtools-extension";
+import thunk from "redux-thunk";
+import cocktailReducer from "../views/Home/reducer";
+
+const RootReducer = combineReducers({
+  cocktails: cocktailReducer,
+});
+
+const store = createStore(
+  RootReducer,
+  composeWithDevTools(applyMiddleware(thunk))
+);
+
+export type RootStore = ReturnType<typeof RootReducer>;
+
+export type AppDispatch = typeof store.dispatch;
+
+export type RootState = ReturnType<typeof store.getState>;
+
+export default store;
