+++ src/views/Home/actions.ts
@@ -1,21 +1,16 @@
+// Cocktails
+
 import axios from "axios";
 import { ThunkDispatch } from "redux-thunk";
-import { RootState, RootStore } from "../../app/store";
+import { RootState } from "../../store";
 import { randomEight } from "../../helpers/functions";
 
 import {
   Cocktail,
   CocktailResponseType,
   ActionType,
-  Cocktails,
   CocktailFetchState,
-} from "./cocktailTypes";
-
-const initialCocktails: Cocktails = {
-  loading: false,
-  error: { hasHappened: false, message: "" },
-  drinks: [],
-};
+} from "./types";
 
 const loadingCocktails = (): {
   type: string;
@@ -58,32 +53,3 @@ export const fetchAlcoholicsDrink = async (
     dispatch(failedCocktails((error as { message: string }).message));
   }
 };
-
-export const cocktailSelector = (state: RootStore): Cocktails =>
-  state.cocktails;
-
-const cocktailReducer = (
-  state: Cocktails = initialCocktails,
-  action: ActionType
-) => {
-  switch (action.type) {
-    case CocktailFetchState.loading:
-      return { ...state, loading: true };
-    case CocktailFetchState.loaded:
-      return {
-        ...state,
-        loading: false,
-        drinks: action.payload as Cocktail[],
-      };
-    case CocktailFetchState.failed:
-      return {
-        ...state,
-        loading: false,
-        error: { hasHappened: true, message: action.payload as string },
-      };
-    default:
-      return state;
-  }
-};
-
-export default cocktailReducer;
